+++ core/moengage_flutter_android/lib/moengage_flutter_android.dart
@@ -307,4 +307,10 @@ class MoEngageFlutterAndroid extends MoEngageFlutterPlatform {
     payload[keyData] = {keyUpdatePushPermissionCount: requestCount};
     return jsonEncode(payload);
   }
+
+  @override
+  void showNudge(MoEngageNudgePosition position, String appId) {
+    _methodChannel.invokeMethod(methodNameShowNudge,
+        jsonEncode(getShowNudgeJsonPayload(position, appId)));
+  }
 }