+++ core/moengage_flutter_platform_interface/lib/src/model/inapp/nudge_position.dart
@@ -1,0 +1,18 @@
+/// Position for Non-Intrusive Nudge Campaigns
+/// @since TODO: Add Version
+enum MoEngageNudgePosition {
+  /// For nudge to be shown at the TOP
+  top,
+
+  /// For nudge to be shown at the BOTTOM
+  bottom,
+
+  /// For nudge to be shown at the BOTTOM_RIGHT
+  bottomRight,
+
+  /// For nudge to be shown at the BOTTOM_LEFT
+  bottomLeft,
+
+  /// For nudge to be shown at any available position
+  any;
+}