+++ core/moengage_flutter_platform_interface/test/parser_test.dart
@@ -61,6 +61,8 @@ void main() {
     expect(
         getUserAttributePayload('number-array', 'general', [1.5, 1, 2.56], ''),
         jsonDecode(userAttrNumberArrayPayload));
+    expect(getShowNudgeJsonPayload(MoEngageNudgePosition.bottom, "1234"),
+        jsonDecode(nudgePayload));
   });
 
   test('Test Alias Payload', () {